refactor(evals): extract CRB_REPOS into a single shared module
config.ts had a slug array and upload-datasets.ts had an independent
slug→language map. Adding a new repo required two edits with no
enforcement they stayed in sync.

Introduces evals/src/crb-repos.ts as the single source of truth.
Both files now import from it; config.ts derives slug lists via
Object.keys(CRB_REPOS).
