+++ evals/src/run-eval.spec.ts
@@ -3,7 +3,8 @@
 jest.mock('langfuse', () => ({ Langfuse: jest.fn() }));
 
 import { parseDiffLines } from './reviewer';
-import { resolveDatasets, resolvePreset, listPresets, CRB_REPOS } from './config';
+import { resolveDatasets, resolvePreset, listPresets } from './config';
+import { CRB_REPOS } from './crb-repos';
 import { classifyError, createRunLog } from './run-log';
 import fs from 'node:fs';
 import path from 'node:path';
